clean up some mess

+++ src/main/java/com/enderio/core/client/ClientUtil.java
@@ -63,7 +63,7 @@ public class ClientUtil {
         }
         try {
             Object val = Y.get(p);
-            return ((Double) val).doubleValue();
+            return (Double) val;
         } catch (Exception e) {
             e.printStackTrace();
             return 0;
